ECB Digital Euro Report Text Mining
This project showcases text mining techniques and sentiment analysis to explore the "Progress on the Investigation Phase of a Digital Euro" report by the European Central Bank (ECB). The primary goal is to extract valuable insights from the report, analyze the sentiments conveyed, and visualize the findings to facilitate better understanding and decision-making.